src/DoanTN-AI/recommendations/requirements.txt
pandas==2.1.0
numpy==1.24.3
scikit-learn==1.3.0
implicit==0.6.2
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
pyodbc==4.0.39
plotly==5.18.0
apscheduler
python-jose[cryptography]
PyJWT
redis
requests